chore: rename cli rules (#1946)
