+++ docs/changelog.md
@@ -244,7 +244,7 @@ toc:
 
 ### Minor Changes
 
-- Added Spot and Arazzo rules: `no-criteria-xpath`, `no-actions-type-end`, `criteria-unique`.
+- Added Respect and Arazzo rules: `no-criteria-xpath`, `no-actions-type-end`, `criteria-unique`.
 
 ### Patch Changes
