+++ docs/rules/spec-ruleset.md
@@ -10,7 +10,7 @@ All the rules are of severity `error`:
 - [struct](./arazzo/struct.md)
 - [criteria-unique](./arazzo/criteria-unique.md)
 - [sourceDescription-type](./arazzo/sourceDescriptions-type.md)
-- [spot-supported-versions](./spot/spot-supported-versions.md)
+- [respect-supported-versions](./respect/respect-supported-versions.md)
 - [workflowId-unique](./arazzo/workflowId-unique.md)
 - [stepId-unique](./arazzo/stepId-unique.md)
 - [sourceDescription-name-unique](./arazzo/sourceDescriptions-name-unique.md)
@@ -20,5 +20,5 @@ All the rules are of severity `error`:
 - [step-onSuccess-unique](./arazzo/step-onSuccess-unique.md)
 - [step-onFailure-unique](./arazzo/step-onFailure-unique.md)
 - [requestBody-replacements-unique](./arazzo/requestBody-replacements-unique.md)
-- [no-criteria-xpath](./spot/no-criteria-xpath.md)
+- [no-criteria-xpath](./respect/no-criteria-xpath.md)
 - [criteria-unique](./arazzo/criteria-unique.md)
