+++ docs/sidebars.yaml
@@ -138,8 +138,8 @@
             - page: rules/async/channels-kebab-case.md
             - page: rules/async/no-channel-trailing-slash.md
             - separator: Arazzo
-            - page: rules/spot/no-criteria-xpath.md
-            - page: rules/spot/spot-supported-versions.md
+            - page: rules/respect/no-criteria-xpath.md
+            - page: rules/respect/respect-supported-versions.md
             - page: rules/arazzo/criteria-unique.md
             - page: rules/arazzo/parameters-unique.md
             - page: rules/arazzo/requestBody-replacements-unique.md
