+++ packages/core/CHANGELOG.md
@@ -32,7 +32,7 @@
 
 ### Patch Changes
 
-- Added support for Arazzo version 1.0.1 in Spot validation rules.
+- Added support for Arazzo version 1.0.1 in Respect validation rules.
 
 ## 1.28.1
 
@@ -73,13 +73,13 @@
 
 ### Patch Changes
 
-- Updated Spot validation rules.
+- Updated Respect validation rules.
 
 ## 1.26.1
 
 ### Patch Changes
 
-- Removed the `no-actions-type-end` Spot rule.
+- Removed the `no-actions-type-end` Respect rule.
 - Removed unused lodash.isequal dependency.
 
 ## 1.26.0
@@ -202,7 +202,7 @@
 
 ### Minor Changes
 
-- Added Spot and Arazzo rules: `no-criteria-xpath`, `no-actions-type-end`, `criteria-unique`.
+- Added Respect and Arazzo rules: `no-criteria-xpath`, `no-actions-type-end`, `criteria-unique`.
 
 ### Patch Changes
