+++ packages/core/src/rules/arazzo/__tests__/no-criteria-xpath.test.ts
@@ -90,7 +90,7 @@ describe('Arazzo no-criteria-xpath', () => {
               "source": "arazzo.yaml",
             },
           ],
-          "message": "The \`xpath\` type criteria is not supported by Spot.",
+          "message": "The \`xpath\` type criteria is not supported by Respect.",
           "ruleId": "no-criteria-xpath",
           "severity": "error",
           "suggest": [],
@@ -103,7 +103,7 @@ describe('Arazzo no-criteria-xpath', () => {
               "source": "arazzo.yaml",
             },
           ],
-          "message": "The \`xpath\` type criteria is not supported by Spot.",
+          "message": "The \`xpath\` type criteria is not supported by Respect.",
           "ruleId": "no-criteria-xpath",
           "severity": "error",
           "suggest": [],
