+++ packages/core/src/rules/respect/respect-supported-versions.ts
@@ -1,20 +1,20 @@
-import { ARAZZO_VERSIONS_SUPPORTED_BY_SPOT } from '../../typings/arazzo';
+import { ARAZZO_VERSIONS_SUPPORTED_BY_RESPECT } from '../../typings/arazzo';
 import { pluralize } from '../../utils';
 
 import type { Arazzo1Rule } from '../../visitors';
 import type { UserContext } from '../../walk';
 
-export const SpotSupportedVersions: Arazzo1Rule = () => {
-  const supportedVersions = ARAZZO_VERSIONS_SUPPORTED_BY_SPOT.join(', ');
+export const RespectSupportedVersions: Arazzo1Rule = () => {
+  const supportedVersions = ARAZZO_VERSIONS_SUPPORTED_BY_RESPECT.join(', ');
   return {
     Root: {
       enter(root, { report, location }: UserContext) {
-        if (!ARAZZO_VERSIONS_SUPPORTED_BY_SPOT.includes(root.arazzo)) {
+        if (!ARAZZO_VERSIONS_SUPPORTED_BY_RESPECT.includes(root.arazzo)) {
           report({
             message: `Only ${supportedVersions} Arazzo ${pluralize(
               'version is',
-              ARAZZO_VERSIONS_SUPPORTED_BY_SPOT.length
-            )} supported by Spot.`,
+              ARAZZO_VERSIONS_SUPPORTED_BY_RESPECT.length
+            )} supported by Respect.`,
             location: location.child('arazzo'),
           });
         }
